+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/statements/pg/AlterView.java
@@ -44,18 +44,19 @@ public class AlterView extends PgParserAbstract {
     @Override
     public void parseObject() {
         List<ParserRuleContext> ids = getIdentifiers(ctx.name);
-        PgView dbView = (PgView) getSafe(AbstractSchema::getView,
-                getSchemaSafe(ids), QNameParser.getFirstNameCtx(ids));
-        Alter_view_actionContext action = ctx.alter_view_action();
-        if (action.set_def_column() != null) {
-            VexContext exp = action.set_def_column().vex();
-            doSafe((s,o) -> {
-                s.addColumnDefaultValue(getFullCtxText(action.column_name), getExpressionText(exp, stream));
-                db.addAnalysisLauncher(new VexAnalysisLauncher(s, exp, fileName));
-            }, dbView, null);
-        }
-        if (action.drop_def() != null) {
-            doSafe(PgView::removeColumnDefaultValue, dbView, getFullCtxText(action.column_name));
+        var st = getSafe(AbstractSchema::getView, getSchemaSafe(ids), QNameParser.getFirstNameCtx(ids));
+        if (st instanceof PgView dbView) {
+            Alter_view_actionContext action = ctx.alter_view_action();
+            if (action.set_def_column() != null) {
+                VexContext exp = action.set_def_column().vex();
+                doSafe((s, o) -> {
+                    s.addColumnDefaultValue(getFullCtxText(action.column_name), getExpressionText(exp, stream));
+                    db.addAnalysisLauncher(new VexAnalysisLauncher(s, exp, fileName));
+                }, dbView, null);
+            }
+            if (action.drop_def() != null) {
+                doSafe(PgView::removeColumnDefaultValue, dbView, getFullCtxText(action.column_name));
+            }
         }
 
         addObjReference(ids, DbObjType.VIEW, ACTION_ALTER);